+++ examples/alerts/alerts.go
@@ -7,22 +7,18 @@ import (
 )
 
 ///////////////////////////////////////////////////
-// Create API keys
-// POST /api_keys
+// Create a new Alert
+// POST /alerts
 
-func CreateAPIkeys() {
+func CreateanewAlert() {
   apiKey := os.Getenv("YOUR_SENDGRID_APIKEY")
   host := "https://api.sendgrid.com"
-  request := sendgrid.GetRequest(apiKey, "/v3/api_keys", host)
+  request := sendgrid.GetRequest(apiKey, "/v3/alerts", host)
   request.Method = "POST"
   request.Body = []byte(` {
-  "name": "My API Key", 
-  "sample": "data", 
-  "scopes": [
-    "mail.send", 
-    "alerts.create", 
-    "alerts.read"
-  ]
+  "email_to": "example@example.com", 
+  "frequency": "daily", 
+  "type": "stats_notification"
 }`)
   response, err := sendgrid.API(request)
   if err != nil {
@@ -35,13 +31,13 @@ func CreateAPIkeys() {
 }
 
 ///////////////////////////////////////////////////
-// Retrieve all API Keys belonging to the authenticated user
-// GET /api_keys
+// Retrieve all alerts
+// GET /alerts
 
-func RetrieveallAPIKeysbelongingtotheauthenticateduser() {
+func Retrieveallalerts() {
   apiKey := os.Getenv("YOUR_SENDGRID_APIKEY")
   host := "https://api.sendgrid.com"
-  request := sendgrid.GetRequest(apiKey, "/v3/api_keys", host)
+  request := sendgrid.GetRequest(apiKey, "/v3/alerts", host)
   request.Method = "GET"
   response, err := sendgrid.API(request)
   if err != nil {
@@ -54,42 +50,16 @@ func RetrieveallAPIKeysbelongingtotheauthenticateduser() {
 }
 
 ///////////////////////////////////////////////////
-// Update the name & scopes of an API Key
-// PUT /api_keys/{api_key_id}
+// Update an alert
+// PATCH /alerts/{alert_id}
 
-func UpdatethenamescopesofanAPIKey() {
+func Updateanalert() {
   apiKey := os.Getenv("YOUR_SENDGRID_APIKEY")
   host := "https://api.sendgrid.com"
-  request := sendgrid.GetRequest(apiKey, "/v3/api_keys/{api_key_id}", host)
-  request.Method = "PUT"
-  request.Body = []byte(` {
-  "name": "A New Hope", 
-  "scopes": [
-    "user.profile.read", 
-    "user.profile.update"
-  ]
-}`)
-  response, err := sendgrid.API(request)
-  if err != nil {
-    fmt.Println(err)
-  } else {
-    fmt.Println(response.StatusCode)
-    fmt.Println(response.Body)
-    fmt.Println(response.Headers)
-  }
-}
-
-///////////////////////////////////////////////////
-// Update API keys
-// PATCH /api_keys/{api_key_id}
-
-func UpdateAPIkeys() {
-  apiKey := os.Getenv("YOUR_SENDGRID_APIKEY")
-  host := "https://api.sendgrid.com"
-  request := sendgrid.GetRequest(apiKey, "/v3/api_keys/{api_key_id}", host)
+  request := sendgrid.GetRequest(apiKey, "/v3/alerts/{alert_id}", host)
   request.Method = "PATCH"
   request.Body = []byte(` {
-  "name": "A New Hope"
+  "email_to": "example@example.com"
 }`)
   response, err := sendgrid.API(request)
   if err != nil {
@@ -102,13 +72,13 @@ func UpdateAPIkeys() {
 }
 
 ///////////////////////////////////////////////////
-// Retrieve an existing API Key
-// GET /api_keys/{api_key_id}
+// Retrieve a specific alert
+// GET /alerts/{alert_id}
 
-func RetrieveanexistingAPIKey() {
+func Retrieveaspecificalert() {
   apiKey := os.Getenv("YOUR_SENDGRID_APIKEY")
   host := "https://api.sendgrid.com"
-  request := sendgrid.GetRequest(apiKey, "/v3/api_keys/{api_key_id}", host)
+  request := sendgrid.GetRequest(apiKey, "/v3/alerts/{alert_id}", host)
   request.Method = "GET"
   response, err := sendgrid.API(request)
   if err != nil {
@@ -121,13 +91,13 @@ func RetrieveanexistingAPIKey() {
 }
 
 ///////////////////////////////////////////////////
-// Delete API keys
-// DELETE /api_keys/{api_key_id}
+// Delete an alert
+// DELETE /alerts/{alert_id}
 
-func DeleteAPIkeys() {
+func Deleteanalert() {
   apiKey := os.Getenv("YOUR_SENDGRID_APIKEY")
   host := "https://api.sendgrid.com"
-  request := sendgrid.GetRequest(apiKey, "/v3/api_keys/{api_key_id}", host)
+  request := sendgrid.GetRequest(apiKey, "/v3/alerts/{alert_id}", host)
   request.Method = "DELETE"
   response, err := sendgrid.API(request)
   if err != nil {